Add PathToArc Parameter on Build Feature & Moving Logging Classes
Add a new field on the build feature dialog allowing the user to specify where arcanist command is located on the agent. We will need this information to execute the patching process upon which we will be building.

This commit also add Command and CommandBuilder into Common which we will be using to execute arcanist to apply the `arc patch` during the build so we can actually build the changes from the revision,

Also moved the three logging based files into a sub-folder causing many file updates of just an import.